merge

+++ common/src/main/java/com/taobao/arthas/common/SocketUtils.java
@@ -17,7 +17,7 @@ public class SocketUtils {
     private SocketUtils() {
     }
 
-    public static int findTcpListenProcess(int port) {
+    public static long findTcpListenProcess(int port) {
         try {
             if (OSUtils.isWindows()) {
                 String[] command = { "netstat", "-ano", "-p", "TCP" };
@@ -28,7 +28,7 @@ public class SocketUtils {
                         String[] strings = line.trim().split("\\s+");
                         if (strings.length == 5) {
                             if (strings[1].endsWith(":" + port)) {
-                                return Integer.parseInt(strings[4]);
+                                return Long.parseLong(strings[4]);
                             }
                         }
                     }
@@ -38,7 +38,7 @@ public class SocketUtils {
             if (OSUtils.isLinux() || OSUtils.isMac()) {
                 String pid = ExecutingCommand.getFirstAnswer("lsof -t -s TCP:LISTEN -i TCP:" + port);
                 if (!pid.trim().isEmpty()) {
-                    return Integer.parseInt(pid);
+                    return Long.parseLong(pid);
                 }
             }
         } catch (Throwable e) {
